Add files via upload
Trained the model using 10 epochs and used a learning rate of 0.0001. Looks like I'm getting a decent accuracy. One main issue is the change in the number of channels. I couldn't figure out what's wrong :( . Every time it passes through a transform the colour channels change so I just changed the input shape to 3 as well. If anyone has a solution to this let me know!